change: Extract LinodeConfigs component.
